requirements.txt
# ============================================
# 多模态文档问答助手 - 依赖清单
# ============================================

# --- Web 框架 ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# --- AI / LLM ---
langchain==1.2.13
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-openai==1.1.12
zai-sdk==0.2.2

# --- 向量检索 ---
faiss-cpu==1.13.2

# --- PDF 处理 ---
PyMuPDF==1.27.2

# --- 图片处理 ---
Pillow==12.1.1

# --- 工具 ---
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
